IT WORKS! First working iteration.
It's slow to scan a UPC; I plan to improve that by limiting location
and orientation within the camera frame rather than using autodetect.

But after it gets through the Musicbrainz rate limiting, it renders
the new and old search systems alongside the video feed for scanning
barcodes. It also reports the album artist and title and date of
publication.

Pretty freaking sweet.